search: Add ExpandRepo
ExpandRepo allows us to find out which repositories need to be searched in
most cases by translating repo tokens in queries into a set of repositories to
be searched.

It still can't give us a final list, due to queries such as `(not (reposet
foo))`. That will require a different function.