search: MinimalRepoSet
MinimalRepoSet is a function to calculate the minimal set of repositories that
apply to a query. Zoekt runs a search against every repository, but first
replaces each repo expression with a constant before continuing the
search. Via constant folding it can determine if it should search the
repository. This is efficient for 10000s of repositories, but does not scale
to Sourcegraph's repo list which is millions of repositories. Additionally we
need to work out the set since searcher has an upper bound on the number of
repos it will search.

We can convert a query expression into an expression of set operations, which
is what this commit implements.

Future work: Investigate converting the expression into a SQL where query. The
original approach for this commit wouldn't have worked, but this current
approach may work.